fix(metrics): fix issue where package download counters were not fully accurate
During real world testing, the count of package downloads almost always came in lower then it should
have. It seems that sometimes the npm client kills the connection prior to Express thinking the full
response has been sent. The solution seems to be to rely on the `close` event instead of the
`finish` event on the response object.

fix #7